+++ src/EntityGraphQL/Schema/Field.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
+using EntityGraphQL.Authorization;
 using EntityGraphQL.Extensions;
 
 namespace EntityGraphQL.Schema