+++ src/EntityGraphQL/Schema/SchemaProvider.cs
@@ -38,14 +38,14 @@ namespace EntityGraphQL.Schema
         public SchemaProvider()
         {
             // default GQL scalar types
-            types.Add("Int", new SchemaType<int>(this, "Int", "Int scalar", false, false, true));
-            types.Add("Float", new SchemaType<double>(this, "Float", "Float scalar", false, false, true));
-            types.Add("Boolean", new SchemaType<bool>(this, "Boolean", "Boolean scalar", false, false, true));
-            types.Add("String", new SchemaType<string>(this, "String", "String scalar", false, false, true));
-            types.Add("ID", new SchemaType<Guid>(this, "ID", "ID scalar", false, false, true));
+            types.Add("Int", new SchemaType<int>(this, "Int", "Int scalar", null, false, false, true));
+            types.Add("Float", new SchemaType<double>(this, "Float", "Float scalar", null, false, false, true));
+            types.Add("Boolean", new SchemaType<bool>(this, "Boolean", "Boolean scalar", null, false, false, true));
+            types.Add("String", new SchemaType<string>(this, "String", "String scalar", null, false, false, true));
+            types.Add("ID", new SchemaType<Guid>(this, "ID", "ID scalar", null, false, false, true));
 
             // default custom scalar for DateTime
-            types.Add("Date", new SchemaType<DateTime>(this, "Date", "Date with time scalar", false, false, true));
+            types.Add("Date", new SchemaType<DateTime>(this, "Date", "Date with time scalar", null, false, false, true));
 
             customTypeMappings = new Dictionary<Type, GqlTypeInfo> {
                 {typeof(short), new GqlTypeInfo(() => Type("Int"), typeof(short))},
@@ -59,7 +59,7 @@ namespace EntityGraphQL.Schema
                 {typeof(bool), new GqlTypeInfo(() => Type("Boolean"), typeof(bool))},
             };
 
-            var queryContext = new SchemaType<TContextType>(this, typeof(TContextType).Name, "Query schema");
+            var queryContext = new SchemaType<TContextType>(this, typeof(TContextType).Name, "Query schema", null);
             queryContextName = queryContext.Name;
             types.Add(queryContext.Name, queryContext);
 
@@ -144,14 +144,21 @@ namespace EntityGraphQL.Schema
         /// <returns></returns>
         public SchemaType<TBaseType> AddType<TBaseType>(string name, string description)
         {
-            var tt = new SchemaType<TBaseType>(this, name, description);
+            var tt = new SchemaType<TBaseType>(this, name, description, null);
+            var attributes = typeof(TBaseType).GetTypeInfo().GetCustomAttributes(typeof(GraphQLAuthorizeAttribute), true).Cast<GraphQLAuthorizeAttribute>();
+            if (attributes.Any())
+                tt.AuthorizeClaims = new RequiredClaims(attributes);
+
             types.Add(name, tt);
             return tt;
         }
 
         public ISchemaType AddType(Type contextType, string name, string description)
         {
-            var tt = new SchemaType<object>(this, contextType, name, description);
+            var tt = new SchemaType<object>(this, contextType, name, description, null);
+            var attributes = contextType.GetTypeInfo().GetCustomAttributes(typeof(GraphQLAuthorizeAttribute), true).Cast<GraphQLAuthorizeAttribute>();
+            if (attributes.Any())
+                tt.AuthorizeClaims = new RequiredClaims(attributes);
             types.Add(name, tt);
             return tt;
         }
@@ -176,7 +183,11 @@ namespace EntityGraphQL.Schema
         /// <returns></returns>
         public SchemaType<TBaseType> AddInputType<TBaseType>(string name, string description)
         {
-            var tt = new SchemaType<TBaseType>(this, name, description, true);
+            var tt = new SchemaType<TBaseType>(this, name, description, null, true);
+            var attributes = typeof(TBaseType).GetTypeInfo().GetCustomAttributes(typeof(GraphQLAuthorizeAttribute), true).Cast<GraphQLAuthorizeAttribute>();
+            if (attributes.Any())
+                tt.AuthorizeClaims = new RequiredClaims(attributes);
+
             types.Add(name, tt);
             return tt;
         }
@@ -438,7 +449,16 @@ namespace EntityGraphQL.Schema
             if (!types.ContainsKey(typeName))
                 throw new EntityQuerySchemaException($"{typeName} not found in schema.");
 
-            var field = types[typeName].GetField(fieldName, claims);
+            ISchemaType schemaType = types[typeName];
+
+            if (!AuthUtil.IsAuthorized(claims, schemaType.AuthorizeClaims))
+                throw new EntityGraphQLAccessException($"You do not have access to the '{typeName}' type. You require any of the following security claims [{string.Join(", ", schemaType.AuthorizeClaims.Claims.SelectMany(r => r))}]");
+
+            var field = schemaType.GetField(fieldName, claims);
+
+            if (!AuthUtil.IsAuthorized(claims, field.ReturnType.SchemaType.AuthorizeClaims))
+                throw new EntityGraphQLAccessException($"You do not have access to the '{field.ReturnType.SchemaType.Name}' type. You require any of the following security claims [{string.Join(", ", field.ReturnType.SchemaType.AuthorizeClaims.Claims.SelectMany(r => r))}]");
+
             var result = new ExpressionResult(field.Resolve ?? Expression.Property(context, fieldName), field.Services);
 
             if (field.ArgumentTypesObject != null)
@@ -664,7 +684,7 @@ namespace EntityGraphQL.Schema
 
         public ISchemaType AddScalarType<TType>(string gqlTypeName, string description)
         {
-            var schemaType = new SchemaType<TType>(this, gqlTypeName, description, false, false, true);
+            var schemaType = new SchemaType<TType>(this, gqlTypeName, description, null, false, false, true);
             types.Add(gqlTypeName, schemaType);
             return schemaType;
         }
@@ -723,7 +743,10 @@ namespace EntityGraphQL.Schema
 
         public ISchemaType AddEnum(string name, Type type, string description)
         {
-            var schemaType = new SchemaType<object>(this, type, name, description, false, true);
+            var schemaType = new SchemaType<object>(this, type, name, description, null, false, true);
+            var attributes = type.GetTypeInfo().GetCustomAttributes(typeof(GraphQLAuthorizeAttribute), true).Cast<GraphQLAuthorizeAttribute>();
+            if (attributes.Any())
+                schemaType.AuthorizeClaims = new RequiredClaims(attributes);
             types.Add(name, schemaType);
             return schemaType.AddAllFields();
         }